# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  clap_internal_modules_realtime_dto.TimeSyncRequest:
    properties:
      client_timestamp_ms:
        type: integer
    required:
    - client_timestamp_ms
    type: object
  clap_internal_modules_season_dto.CreateSeasonRequest:
    properties:
      end_date:
        type: string
      is_active:
        type: boolean
      league_id:
        type: string
      name:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - league_id
    - name
    - start_date
    type: object
  clap_internal_modules_season_dto.UpdateSeasonRequest:
    properties:
      end_date:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - name
    - start_date
    type: object
...
  clap_internal_shared_response.ErrorDetail:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  clap_internal_shared_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/clap_internal_shared_response.ErrorDetail'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: Clap API Support
  description: Clap match-day experience backend (auth, clubs, matches, realtime,
    shop).
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: Clap Backend API
  version: "1.0"
paths:
...
  /api/v1/leagues/{id}/seasons:
    get:
      parameters:
      - description: League ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: List seasons by league
      tags:
      - seasons
...
  /api/v1/realtime/time:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: Server time
      tags:
      - realtime
  /api/v1/realtime/time-sync:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clap_internal_modules_realtime_dto.TimeSyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: Time sync
      tags:
      - realtime
  /api/v1/realtime/ws:
    get:
      description: Upgrade to WebSocket. JWT required (query or header).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: WebSocket connect
      tags:
      - realtime
  /api/v1/seasons:
    get:
      parameters:
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: asc|desc
        in: query
        name: order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: List seasons
      tags:
      - seasons
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clap_internal_modules_season_dto.CreateSeasonRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Create season
      tags:
      - seasons
  /api/v1/seasons/{id}:
    delete:
      parameters:
      - description: Season ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Delete season
      tags:
      - seasons
    get:
      parameters:
      - description: Season ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      summary: Get season
      tags:
      - seasons
    put:
      consumes:
      - application/json
      parameters:
      - description: Season ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clap_internal_modules_season_dto.UpdateSeasonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update season
      tags:
      - seasons
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"